Lots of minor changes
1. Removed commands
 scale_rd : not properly functioning. No one used it, so it was removed.
 wpi, wpi_grid, fvm : never implemented

2. polar_damp_type none option was added. Same as polar_damp_type off.
 Also fixed bug that required polar_damp <double> be set even when damp_type = none/off

3. Removed old sample configurations. added new ones. Will probably update old sample
 configs later and add comments/documentation, but no promises. Also, please do not
 add configs later without proper comments and documentation.

4. fixed bug in polar_ewald (polar_ewald.c:ewald_estatic()). Coincidentally, it is
 now evident that there is a bug in polar_ewald_full as it's systematically less
 accurate than polar_ewald.

5. added polar_ewald_alpha command for manually overriding the default value of
 3.5/volume (though this default should be fine in most cases). Also, modified
 the code to allow for manually setting polar_ewald_alpha and ewald_alpha in
 non-constant volume ensembles (e.g. replay and NPT)

6. fixed memory leak in cleanup.c:free_thole_matrices() when run in total_energy/replay ensembles.

7. fixed incorrect datatype for return value for usefulmath.c:iidotprod()

+++ include/function_prototypes.h
@@ -249,4 +249,4 @@ double co2_fugacity(double, double);
 //useful math calls
 double dddotprod(double*, double*);
 double didotprod(double*, int*);
-double iidotprod(int*, int*);
+int iidotprod(int*, int*);
